feat(earthquake_notify): implement dynamic command visibility
