--- src/localization/en-US/commands/misc/earthquake_notify.jsonc
@@ -1,39 +1,0 @@
-{
-    "$schema": "../../../../types/structure/schemas/command.json",
-    "earthquake_notify": {
-        "name": "earthquake_notify",
-        "pretty_name": "Earthquake Notification",
-        "description": "Manage your earthquake notification subscriptions.",
-        "help": "Subscribe to receive direct messages when an earthquake occurs in specific cities.\n\n**Usage:**\n- `/earthquake_notify add <city>` - Subscribe to a city.\n- `/earthquake_notify remove <city>` - Unsubscribe from a city.\n- `/earthquake_notify list` - List your subscriptions.",
-        "parameters": {
-            "subcommand": {
-                "name": "action",
-                "description": "The action to perform (add, remove, list).",
-                "options": {
-                    "add": "Add Subscription",
-                    "remove": "Remove Subscription",
-                    "list": "List Subscriptions",
-                },
-            },
-            "city": {
-                "name": "city",
-                "description": "The name of the city to subscribe to.",
-            },
-            "user": {
-                "name": "user",
-                "description": "The user to manage subscriptions for (Admin only).",
-            },
-        },
-        "execute": {
-            "added": "Successfully subscribed to earthquake notifications for **{city}**.",
-            "removed": "Successfully unsubscribed from earthquake notifications for **{city}**.",
-            "not_found": "No subscription found for **{city}**.",
-            "already_exists": "You are already subscribed to **{city}**.",
-            "list_header": "Earthquake Subscriptions for {user}",
-            "no_subscriptions": "No active subscriptions found.",
-            "admin_only": "You do not have permission to manage subscriptions for other users.",
-            "city_not_found": "The city **{city}** was not found or is not recognized as a valid city.",
-            "api_error": "An error occurred while validating the city. Please try again later.",
-        },
-    },
-}
