--- src/localization/tr/commands/misc/earthquake_notify.jsonc
@@ -1,39 +1,0 @@
-{
-    "$schema": "../../../../types/structure/schemas/command.json",
-    "earthquake_notify": {
-        "name": "earthquake_notify",
-        "pretty_name": "Deprem Bildirimi",
-        "description": "Deprem bildirim aboneliklerinizi yönetin.",
-        "help": "Belirli şehirlerde deprem olduğunda direkt mesaj almak için abone olun.\n\n**Kullanım:**\n- `/earthquake_notify add <şehir>` - Bir şehre abone ol.\n- `/earthquake_notify remove <şehir>` - Bir şehirden aboneliği kaldır.\n- `/earthquake_notify list` - Aboneliklerini listele.",
-        "parameters": {
-            "subcommand": {
-                "name": "işlem",
-                "description": "Yapılacak işlem (ekle, kaldır, listele).",
-                "options": {
-                    "add": "Abonelik Ekle",
-                    "remove": "Aboneliği Kaldır",
-                    "list": "Abonelikleri Listele",
-                },
-            },
-            "city": {
-                "name": "şehir",
-                "description": "Abone olunacak şehrin adı.",
-            },
-            "user": {
-                "name": "kullanıcı",
-                "description": "Abonelikleri yönetilecek kullanıcı (Sadece Yönetici).",
-            },
-        },
-        "execute": {
-            "added": "**{city}** için deprem bildirimlerine başarıyla abone olundu.",
-            "removed": "**{city}** için deprem bildirimlerinden başarıyla çıkıldı.",
-            "not_found": "**{city}** için abonelik bulunamadı.",
-            "already_exists": "Zaten **{city}** için abonesiniz.",
-            "list_header": "{user} için Deprem Abonelikleri",
-            "no_subscriptions": "Aktif abonelik bulunamadı.",
-            "admin_only": "Diğer kullanıcıların aboneliklerini yönetmek için yetkiniz yok.",
-            "city_not_found": "**{city}** şehri bulunamadı veya geçerli bir şehir olarak tanınmadı.",
-            "api_error": "Şehir doğrulanırken bir hata oluştu. Lütfen daha sonra tekrar deneyin.",
-        },
-    },
-}
